PR (41): Pass processor
https://github.com/128technology/telegraf/pull/41
- 4863386af: Correctly drop the metrics
- 1069ca085: Renamed processor from filter to pass
